[WPE] Rename WPEView as WPEWebView https://bugs.webkit.org/show_bug.cgi?id=265191
Reviewed by Don Olmstead.

WPEView will conflict with new WPE platform API.

* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/UIProcess/API/C/wpe/WKView.cpp:
* Source/WebKit/UIProcess/API/glib/WebKitUserContentManager.cpp:
* Source/WebKit/UIProcess/API/glib/WebKitWebView.cpp:
* Source/WebKit/UIProcess/API/wpe/PageClientImpl.cpp:
* Source/WebKit/UIProcess/API/wpe/WPEWebView.cpp: Renamed from Source/WebKit/UIProcess/API/wpe/WPEView.cpp.
(WKWPE::View::View):
(WKWPE::m_backend):
(WKWPE::View::~View):
(WKWPE::View::setClient):
(WKWPE::View::frameDisplayed):
(WKWPE::View::willStartLoad):
(WKWPE::View::didChangePageID):
(WKWPE::View::didReceiveUserMessage):
(WKWPE::View::webResourceLoadManager):
(WKWPE::View::setInputMethodContext):
(WKWPE::View::inputMethodContext const):
(WKWPE::View::setInputMethodState):
(WKWPE::View::selectionDidChange):
(WKWPE::View::setSize):
(WKWPE::View::setViewState):
(WKWPE::View::handleKeyboardEvent):
(WKWPE::View::synthesizeCompositionKeyPress):
(WKWPE::View::close):
(WKWPE::View::setFullScreen):
(WKWPE::View::accessible const):
(WKWPE::View::platformWebPageProxyForGamepadInput):
* Source/WebKit/UIProcess/API/wpe/WPEWebView.h: Renamed from Source/WebKit/UIProcess/API/wpe/WPEView.h.
(WKWPE::View::create):
(WKWPE::View::page):
(WKWPE::View::client const):
(WKWPE::View::backend):
(WKWPE::View::size const):
(WKWPE::View::viewState const):
(WKWPE::View::isFullScreen):
(WKWPE::View::touchGestureController const):
* Source/WebKit/UIProcess/API/wpe/WebKitPopupMenu.cpp:
* Source/WebKit/UIProcess/Gamepad/libwpe/UIGamepadProviderLibWPE.cpp:

Canonical link: https://commits.webkit.org/271034@main
