+++ Source/WebKit/SourcesWPE.txt
@@ -207,7 +207,7 @@ UIProcess/API/wpe/WebKitWebViewBackend.cpp @no-unify
 UIProcess/API/wpe/WebKitWebViewWPE1.cpp @no-unify
 UIProcess/API/wpe/WebKitWebViewWPE2.cpp @no-unify
 UIProcess/API/wpe/WebKitWebViewWPE.cpp @no-unify
-UIProcess/API/wpe/WPEView.cpp @no-unify
+UIProcess/API/wpe/WPEWebView.cpp @no-unify
 
 UIProcess/Automation/cairo/WebAutomationSessionCairo.cpp
