+++ Source/WebKit/UIProcess/API/wpe/PageClientImpl.cpp
@@ -32,7 +32,7 @@
 #include "NativeWebTouchEvent.h"
 #include "NativeWebWheelEvent.h"
 #include "TouchGestureController.h"
-#include "WPEView.h"
+#include "WPEWebView.h"
 #include "WebContextMenuProxy.h"
 #include "WebContextMenuProxyWPE.h"
 #include "WebKitPopupMenu.h"
